+++ src/yasbd/__init__.py
@@ -2,6 +2,7 @@ from importlib.metadata import PackageNotFoundError, version
 
 from yasbd.boundary_detector import BoundaryDetector
 from yasbd.boundary_detector import ParagraphEOF
+from yasbd.exceptions import CleanStepError
 from yasbd.exceptions import InvalidInputError
 from yasbd.exceptions import UnsupportedLanguageError
 from yasbd.exceptions import YasbdError
@@ -15,6 +16,7 @@ except PackageNotFoundError:
 __all__ = [
     "BoundaryDetector",
     "ParagraphEOF",
+    "CleanStepError",
     "InvalidInputError",
     "UnsupportedLanguageError",
     "YasbdError",